pyproject.toml
[tool.poetry]
name = "fastapi-starter"
version = "0.1.0"
description = ""
authors = ["MahmudulHassan5809 <mahmudul.hassan240@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.3"}
pydantic-settings = "^2.6.0"
redis = "^5.2.0"
stackprinter = "^0.2.12"
sqlalchemy = "^2.0.36"
bcrypt = "^4.2.0"
dependency-injector = "^4.42.0"
requests = "^2.32.3"
python-ulid = "^3.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
black = "^24.10.0"
mypy = "^1.13.0"
pylint = "^3.3.1"
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
